refactor: move generators and factory to application layer
CRITICAL architectural fixes following Clean Architecture principles:

1. **Moved AI Generators to Application Layer**
   - Relocated 5 generators from domain/services to application/generators:
     * task_generator.go
     * dev_notes_generator.go
     * testing_generator.go
     * scenarios_generator.go
     * qa_assessment_generator.go
   - Reason: Generators are APPLICATION orchestrators, not domain services
   - They coordinate AI calls and infrastructure, not business logic

2. **Moved StoryFactory to Application Layer**
   - Relocated story_factory.go from domain/services to application/factories
   - Reason: Factories orchestrate workflows (application concern), not domain logic

3. **Moved AIClient Port to Domain Ports**
   - Relocated ai_client.go to domain/ports/ai_client_port.go
   - Reason: Port interfaces belong in domain layer (dependency inversion)

4. **Updated All References**
   - Updated package declarations (services → generators, factories, ports)
   - Fixed all imports throughout codebase
   - Updated app/container.go and commands to use new locations

Architecture compliance:
- Domain: Business logic, entities, ports (interfaces)
- Application: Use cases, orchestration, factories, generators
- Infrastructure: External adapters (git, config, templates)
- Adapters: Implementation of ports (AI client, GitHub client)

Build verified passing.
