+++ scripts/bmad-cli/internal/application/generators/qa_assessment_generator.go
@@ -1,6 +1,7 @@
-package services
+package generators
 
 import (
+	"bmad-cli/internal/domain/ports"
 	"bmad-cli/internal/common/ai"
 	"context"
 	"fmt"
@@ -16,7 +17,7 @@ import (
 
 // AIQAAssessmentGenerator generates QA results for stories using AI
 type AIQAAssessmentGenerator struct {
-	aiClient AIClient
+	aiClient ports.AIClient
 	config   *config.ViperConfig
 }
 
@@ -29,7 +30,7 @@ type QAAssessmentData struct {
 }
 
 // NewAIQAAssessmentGenerator creates a new QA assessment generator
-func NewAIQAAssessmentGenerator(aiClient AIClient, config *config.ViperConfig) *AIQAAssessmentGenerator {
+func NewAIQAAssessmentGenerator(aiClient ports.AIClient, config *config.ViperConfig) *AIQAAssessmentGenerator {
 	return &AIQAAssessmentGenerator{
 		aiClient: aiClient,
 		config:   config,
