+++ scripts/bmad-cli/internal/domain/ports/ai_client_port.go
@@ -1,4 +1,4 @@
-package services
+package ports
 
 import (
 	"context"
